+++ src/main/bootstrap/try_peer.rs
@@ -15,8 +15,8 @@
 // Please review the Licences for the specific language governing permissions and limitations
 // relating to use of the SAFE Network Software.
 
-
-use common::{Core, Message, NameHash, Priority, Socket, State};
+use common::{BootstrapDenyReason, Core, ExternalReachability, Message, NameHash, Priority, Socket,
+             State};
 use main::PeerId;
 use mio::{Poll, PollOpt, Ready, Token};
 use rust_sodium::crypto::box_::PublicKey;
@@ -26,13 +26,11 @@ use std::mem;
 use std::net::SocketAddr;
 use std::rc::Rc;
 
-// TODO(Spandan) Result contains socket address too as currently due to bug in mio we are unable to
-// obtain peer address from a connected socket. Track https://github.com/carllerche/mio/issues/397
-// and remove this once that is solved.
 pub type Finish = Box<FnMut(&mut Core,
                             &Poll,
                             Token,
-                            Result<(Socket, SocketAddr, PeerId), SocketAddr>)>;
+                            Result<(Socket, SocketAddr, PeerId),
+                                   (SocketAddr, Option<BootstrapDenyReason>)>)>;
 
 pub struct TryPeer {
     token: Token,
@@ -48,6 +46,7 @@ impl TryPeer {
                  peer: SocketAddr,
                  our_pk: PublicKey,
                  name_hash: NameHash,
+                 ext_reachability: ExternalReachability,
                  finish: Finish)
                  -> ::Res<Token> {
         let socket = Socket::connect(&peer)?;
@@ -62,7 +61,7 @@ impl TryPeer {
             token: token,
             peer: peer,
             socket: socket,
-            request: Some((Message::BootstrapRequest(our_pk, name_hash), 0)),
+            request: Some((Message::BootstrapRequest(our_pk, name_hash, ext_reachability), 0)),
             finish: finish,
         };
 
@@ -73,43 +72,46 @@ impl TryPeer {
 
     fn write(&mut self, core: &mut Core, poll: &Poll, msg: Option<(Message, Priority)>) {
         if self.socket.write(poll, self.token, msg).is_err() {
-            self.handle_error(core, poll);
+            self.handle_error(core, poll, None);
         }
     }
 
-    fn receive_response(&mut self, core: &mut Core, poll: &Poll) {
+    fn read(&mut self, core: &mut Core, poll: &Poll) {
         match self.socket.read::<Message>() {
-            Ok(Some(Message::BootstrapResponse(peer_pk))) => {
+            Ok(Some(Message::BootstrapGranted(peer_pk))) => {
                 let _ = core.remove_state(self.token);
                 let token = self.token;
                 let socket = mem::replace(&mut self.socket, Socket::default());
                 let data = (socket, self.peer, PeerId(peer_pk));
                 (*self.finish)(core, poll, token, Ok(data));
             }
+            Ok(Some(Message::BootstrapDenied(reason))) => {
+                self.handle_error(core, poll, Some(reason))
+            }
             Ok(None) => (),
-            Ok(Some(_)) | Err(_) => self.handle_error(core, poll),
+            Ok(Some(_)) | Err(_) => self.handle_error(core, poll, None),
         }
     }
 
-    fn handle_error(&mut self, core: &mut Core, poll: &Poll) {
+    fn handle_error(&mut self, core: &mut Core, poll: &Poll, reason: Option<BootstrapDenyReason>) {
         self.terminate(core, poll);
         let token = self.token;
         let peer = self.peer;
-        (*self.finish)(core, poll, token, Err(peer));
+        (*self.finish)(core, poll, token, Err((peer, reason)));
     }
 }
 
 impl State for TryPeer {
     fn ready(&mut self, core: &mut Core, poll: &Poll, kind: Ready) {
         if kind.is_error() || kind.is_hup() {
-            self.handle_error(core, poll);
+            self.handle_error(core, poll, None);
         } else {
             if kind.is_writable() {
                 let req = self.request.take();
                 self.write(core, poll, req);
             }
             if kind.is_readable() {
-                self.receive_response(core, poll)
+                self.read(core, poll)
             }
         }
     }