+++ src/main/connection_listener/exchange_msg.rs
@@ -16,16 +16,20 @@
 // relating to use of the SAFE Network Software.
 
 
-use common::{self, Core, CoreTimer, Message, NameHash, Priority, Socket, State};
+use super::check_reachability::CheckReachability;
+use common::{self, BootstrapDenyReason, Core, CoreTimer, ExternalReachability, Message, NameHash,
+             Priority, Socket, State};
 use main::{ActiveConnection, ConnectionCandidate, ConnectionId, ConnectionMap, Event, PeerId};
 use mio::{Poll, PollOpt, Ready, Token};
 use mio::timer::Timeout;
+use nat::ip_addr_is_global;
 use rust_sodium::crypto::box_::PublicKey;
 use std::any::Any;
 use std::cell::RefCell;
+use std::collections::HashSet;
 use std::collections::hash_map::Entry;
 use std::mem;
-use std::rc::Rc;
+use std::rc::{Rc, Weak};
 use std::time::Duration;
 
 pub const EXCHANGE_MSG_TIMEOUT_SEC: u64 = 10 * 60;
@@ -39,6 +43,8 @@ pub struct ExchangeMsg {
     our_pk: PublicKey,
     socket: Socket,
     timeout: Timeout,
+    reachability_children: HashSet<Token>,
+    self_weak: Weak<RefCell<ExchangeMsg>>,
 }
 
 impl ExchangeMsg {
@@ -60,7 +66,7 @@ impl ExchangeMsg {
             core.set_timeout(Duration::from_secs(timeout_sec.unwrap_or(EXCHANGE_MSG_TIMEOUT_SEC)),
                              CoreTimer::new(token, 0))?;
 
-        let state = ExchangeMsg {
+        let state = Rc::new(RefCell::new(ExchangeMsg {
             token: token,
             cm: cm,
             event_tx: event_tx,
@@ -69,20 +75,32 @@ impl ExchangeMsg {
             our_pk: our_pk,
             socket: socket,
             timeout: timeout,
-        };
+            reachability_children: HashSet::with_capacity(4),
+            self_weak: Default::default(),
+        }));
 
-        let _ = core.insert_state(token, Rc::new(RefCell::new(state)));
+        state.borrow_mut().self_weak = Rc::downgrade(&state);
+
+        let _ = core.insert_state(token, state);
 
         Ok(())
     }
 
     fn read(&mut self, core: &mut Core, poll: &Poll) {
         match self.socket.read::<Message>() {
-            Ok(Some(Message::BootstrapRequest(their_public_key, name_hash))) => {
-                self.handle_bootstrap_req(core, poll, their_public_key, name_hash)
+            Ok(Some(Message::BootstrapRequest(their_public_key, name_hash, ext_reachability))) => {
+                match self.get_peer_id(their_public_key) {
+                    Ok(their_id) => {
+                        self.handle_bootstrap_req(core, poll, their_id, name_hash, ext_reachability)
+                    }
+                    Err(()) => self.terminate(core, poll),
+                }
             }
             Ok(Some(Message::Connect(their_public_key, name_hash))) => {
-                self.handle_connect(core, poll, their_public_key, name_hash)
+                match self.get_peer_id(their_public_key) {
+                    Ok(their_id) => self.handle_connect(core, poll, their_id, name_hash),
+                    Err(()) => self.terminate(core, poll),
+                }
             }
             Ok(Some(Message::EchoAddrReq)) => self.handle_echo_addr_req(core, poll),
             Ok(Some(message)) => {
@@ -100,27 +118,88 @@ impl ExchangeMsg {
     fn handle_bootstrap_req(&mut self,
                             core: &mut Core,
                             poll: &Poll,
-                            their_public_key: PublicKey,
-                            name_hash: NameHash) {
-        let their_id = match self.get_peer_id(their_public_key, name_hash) {
-            Ok(their_id) => their_id,
-            Err(()) => return self.terminate(core, poll),
-        };
+                            their_id: PeerId,
+                            name_hash: NameHash,
+                            ext_reachability: ExternalReachability) {
+        if !self.is_valid_name_hash(name_hash) {
+            trace!("Rejecting Bootstrapper with an invalid name hash.");
+            return self.write(core,
+                       poll,
+                       Some((Message::BootstrapDenied(BootstrapDenyReason::InvalidNameHash), 0)));
+        }
+        match ext_reachability {
+            ExternalReachability::Required { direct_listeners } => {
+                for their_listener in direct_listeners.into_iter()
+                    .filter(|addr| ip_addr_is_global(&addr.ip())) {
+                    let self_weak = self.self_weak.clone();
+                    let finish = move |core: &mut Core, poll: &Poll, child, res| {
+                        if let Some(self_rc) = self_weak.upgrade() {
+                            self_rc.borrow_mut().handle_check_reachability(core, poll, child, res)
+                        }
+                    };
+
+                    if let Ok(child) = CheckReachability::<PeerId>::start(core,
+                                                                          poll,
+                                                                          their_listener,
+                                                                          their_id,
+                                                                          Box::new(finish)) {
+                        let _ = self.reachability_children.insert(child);
+                    }
+                }
+                if self.reachability_children.is_empty() {
+                    trace!("Bootstrapper failed to pass requisite condition of external \
+                            recheability. Denying bootstrap.");
+                    let reason = BootstrapDenyReason::FailedExternalReachability;
+                    self.write(core, poll, Some((Message::BootstrapDenied(reason), 0)));
+                }
+            }
+            ExternalReachability::NotRequired => self.send_bootstrap_resp(core, poll, their_id),
+        }
+    }
+
+    fn handle_check_reachability(&mut self,
+                                 core: &mut Core,
+                                 poll: &Poll,
+                                 child: Token,
+                                 res: Result<PeerId, ()>) {
+        let _ = self.reachability_children.remove(&child);
+        if let Ok(their_id) = res {
+            for child in self.reachability_children.drain() {
+                let child = match core.get_state(child) {
+                    Some(state) => state,
+                    None => continue,
+                };
+
+                child.borrow_mut().terminate(core, poll);
+            }
+            return self.send_bootstrap_resp(core, poll, their_id);
+        }
+        if self.reachability_children.is_empty() {
+            trace!("Bootstrapper failed to pass requisite condition of external recheability. \
+                    Denying bootstrap.");
+            let reason = BootstrapDenyReason::FailedExternalReachability;
+            self.write(core, poll, Some((Message::BootstrapDenied(reason), 0)));
+        }
+    }
+
+    fn send_bootstrap_resp(&mut self, core: &mut Core, poll: &Poll, their_id: PeerId) {
+        self.enter_handshaking_mode(their_id);
 
         let our_pk = self.our_pk;
         self.next_state = NextState::ActiveConnection(their_id);
-        self.write(core, poll, Some((Message::BootstrapResponse(our_pk), 0)))
+        self.write(core, poll, Some((Message::BootstrapGranted(our_pk), 0)))
     }
 
     fn handle_connect(&mut self,
                       core: &mut Core,
                       poll: &Poll,
-                      their_public_key: PublicKey,
+                      their_id: PeerId,
                       name_hash: NameHash) {
-        let their_id = match self.get_peer_id(their_public_key, name_hash) {
-            Ok(their_id) => their_id,
-            Err(()) => return self.terminate(core, poll),
-        };
+        if !self.is_valid_name_hash(name_hash) {
+            return self.terminate(core, poll);
+        }
+
+        self.enter_handshaking_mode(their_id);
 
         let our_pk = self.our_pk;
         let name_hash = self.name_hash;
@@ -139,28 +218,30 @@ impl ExchangeMsg {
         }
     }
 
-    fn get_peer_id(&self, their_public_key: PublicKey, name_hash: NameHash) -> Result<PeerId, ()> {
-        if self.our_pk == their_public_key {
-            warn!("Accepted connection from ourselves");
-            return Err(());
-        }
+    fn enter_handshaking_mode(&self, their_id: PeerId) {
+        let mut guard = unwrap!(self.cm.lock());
+        guard.entry(their_id)
+            .or_insert(ConnectionId {
+                active_connection: None,
+                currently_handshaking: 0,
+            })
+            .currently_handshaking += 1;
+        trace!("Connection Map inserted: {:?} -> {:?}",
+               their_id,
+               guard.get(&their_id));
+    }
 
-        if self.name_hash != name_hash {
-            warn!("Incompatible protocol version");
+    fn is_valid_name_hash(&self, name_hash: NameHash) -> bool {
+        self.name_hash == name_hash
+    }
+
+    fn get_peer_id(&self, their_public_key: PublicKey) -> Result<PeerId, ()> {
+        if self.our_pk == their_public_key {
+            debug!("Accepted connection from ourselves");
             return Err(());
         }
 
         let their_id = PeerId(their_public_key);
-
-        {
-            let mut guard = unwrap!(self.cm.lock());
-            guard.entry(their_id)
-                .or_insert(ConnectionId {
-                    active_connection: None,
-                    currently_handshaking: 0,
-                })
-                .currently_handshaking += 1;
-        }
 
         Ok(their_id)
     }
@@ -181,7 +262,7 @@ impl ExchangeMsg {
             Ok(true) => self.done(core, poll),
             Ok(false) => (),
             Err(e) => {
-                warn!("Error in writting: {:?}", e);
+                debug!("Error in writting: {:?}", e);
                 self.terminate(core, poll)
             }
         }
@@ -264,6 +345,9 @@ impl State for ExchangeMsg {
                         let _ = oe.remove();
                     }
                 }
+                trace!("Connection Map removed: {:?} -> {:?}",
+                       their_id,
+                       guard.get(&their_id));
             }
             NextState::None => (),
         }