+++ src/main/service.rs
@@ -15,7 +15,11 @@
 // Please review the Licences for the specific language governing permissions and limitations
 // relating to use of the SAFE Network Software.
 
-use common::{self, Core, CoreMessage, EventLoop, NameHash, Priority};
+// FIXME combine common::self with others if rustfmt does not choke - right now it chokes with
+// v0.7.1
+
+use common;
+use common::{Core, CoreMessage, CrustUser, EventLoop, ExternalReachability, NameHash, Priority};
 use main::{ActiveConnection, Bootstrap, Connect, ConnectionId, ConnectionInfoResult,
            ConnectionListener, ConnectionMap, CrustError, Event, PeerId, PrivConnectionInfo,
            PubConnectionInfo};
@@ -37,7 +41,10 @@ const SERVICE_DISCOVERY_TOKEN: Token = Token(1);
 const LISTENER_TOKEN: Token = Token(2);
 
 const SERVICE_DISCOVERY_DEFAULT_PORT: u16 = 5484;
-/// A structure representing a connection manager. This is the main object through which crust is
+
+const DISABLE_NAT: bool = true;
+
+/// A structure representing all the Crust services. This is the main object through which crust is
 /// used.
 pub struct Service {
     config: Config,
@@ -101,7 +108,7 @@ impl Service {
                                                         our_listeners,
                                                         SERVICE_DISCOVERY_TOKEN,
                                                         port) {
-                    warn!("Could not start ServiceDiscovery: {:?}", e);
+                    debug!("Could not start ServiceDiscovery: {:?}", e);
                 }
             }
             () // Only to get rustfmt happy else it corrects it in a way it detects error
@@ -185,6 +192,22 @@ impl Service {
         }
     }
 
+    /// Returns whether the given peer's IP is in the config file's hard-coded contacts list.
+    pub fn is_peer_hard_coded(&self, peer_id: &PeerId) -> bool {
+        match self.get_peer_socket_addr(peer_id) {
+            Ok(ip) => {
+                debug!("Checking whether {:?} is hard-coded in {:?}",
+                       ip,
+                       self.config.hard_coded_contacts);
+                self.config.hard_coded_contacts.iter().any(|addr| addr.0.ip() == ip.0.ip())
+            }
+            Err(e) => {
+                debug!("{}", e.description());
+                false
+            }
+        }
+    }
+
     // TODO temp remove
     /// Check if we have peers on LAN
     pub fn has_peers_on_lan(&self) -> bool {
@@ -215,17 +238,32 @@ impl Service {
 
     /// Start the bootstrapping procedure. It will auto terminate after indicating success or
     /// failure via the event channel.
-    pub fn start_bootstrap(&mut self, blacklist: HashSet<SocketAddr>) -> ::Res<()> {
+    pub fn start_bootstrap(&mut self,
+                           blacklist: HashSet<SocketAddr>,
+                           crust_user: CrustUser)
+                           -> ::Res<()> {
         let config = self.config.clone();
         let our_pk = self.our_keys.0;
         let name_hash = self.name_hash;
         let cm = self.cm.clone();
         let event_tx = self.event_tx.clone();
+        let ext_reachability = match crust_user {
+            CrustUser::Node => {
+                ExternalReachability::Required {
+                    direct_listeners: unwrap!(self.our_listeners.lock())
+                        .iter()
+                        .map(|s| common::SocketAddr(*s))
+                        .collect(),
+                }
+            }
+            CrustUser::Client => ExternalReachability::NotRequired,
+        };
 
         self.post(move |core, poll| if core.get_state(BOOTSTRAP_TOKEN).is_none() {
             if let Err(e) = Bootstrap::start(core,
                                              poll,
                                              name_hash,
+                                             ext_reachability,
                                              our_pk,
                                              cm,
                                              &config,
@@ -332,44 +370,57 @@ impl Service {
     /// peer, see `Service::connect` for more info.
     // TODO: immediate return in case of sender.send() returned with NotificationError
     pub fn prepare_connection_info(&self, result_token: u32) {
-        let event_tx = self.event_tx.clone();
-        let our_pub_key = self.our_keys.0;
         let our_listeners =
             unwrap!(self.our_listeners.lock()).iter().map(|e| common::SocketAddr(*e)).collect();
-        let mc = self.mc.clone();
-        if let Err(e) = self.post(move |core, poll| {
-            let event_tx_clone = event_tx.clone();
-            match MappedTcpSocket::start(core, poll, 0, &mc, move |_, _, socket, addrs| {
-                let hole_punch_addrs = addrs.into_iter()
-                    .filter(|elt| nat::ip_addr_is_global(&elt.ip()))
-                    .map(common::SocketAddr)
-                    .collect();
-                let event_tx = event_tx_clone;
-                let event = Event::ConnectionInfoPrepared(ConnectionInfoResult {
-                    result_token: result_token,
-                    result: Ok(PrivConnectionInfo {
-                        id: PeerId(our_pub_key),
-                        for_direct: our_listeners,
-                        for_hole_punch: hole_punch_addrs,
-                        hole_punch_socket: socket,
-                    }),
-                });
-                let _ = event_tx.send(event);
-            }) {
-                Ok(()) => (),
-                Err(e) => {
-                    debug!("Error mapping tcp socket: {}", e);
-                    let _ = event_tx.send(Event::ConnectionInfoPrepared(ConnectionInfoResult {
-                        result_token: result_token,
-                        result: Err(From::from(e)),
-                    }));
-                }
-            };
-        }) {
-            let _ = self.event_tx.send(Event::ConnectionInfoPrepared(ConnectionInfoResult {
+        if DISABLE_NAT {
+            let event = Event::ConnectionInfoPrepared(ConnectionInfoResult {
                 result_token: result_token,
-                result: Err(From::from(e)),
-            }));
+                result: Ok(PrivConnectionInfo {
+                    id: PeerId(self.our_keys.0),
+                    for_direct: our_listeners,
+                    for_hole_punch: Default::default(),
+                    hole_punch_socket: None,
+                }),
+            });
+            let _ = self.event_tx.send(event);
+        } else {
+            let event_tx = self.event_tx.clone();
+            let our_pub_key = self.our_keys.0;
+            let mc = self.mc.clone();
+            if let Err(e) = self.post(move |mut core, poll| {
+                let event_tx_clone = event_tx.clone();
+                match MappedTcpSocket::start(core, poll, 0, &mc, move |_, _, socket, addrs| {
+                    let hole_punch_addrs = addrs.into_iter()
+                        .filter(|elt| nat::ip_addr_is_global(&elt.ip()))
+                        .map(common::SocketAddr)
+                        .collect();
+                    let event_tx = event_tx_clone;
+                    let event = Event::ConnectionInfoPrepared(ConnectionInfoResult {
+                        result_token: result_token,
+                        result: Ok(PrivConnectionInfo {
+                            id: PeerId(our_pub_key),
+                            for_direct: our_listeners,
+                            for_hole_punch: hole_punch_addrs,
+                            hole_punch_socket: Some(socket),
+                        }),
+                    });
+                    let _ = event_tx.send(event);
+                }) {
+                    Ok(()) => (),
+                    Err(e) => {
+                        debug!("Error mapping tcp socket: {}", e);
+                        let _ = event_tx.send(Event::ConnectionInfoPrepared(ConnectionInfoResult {
+                                result_token: result_token,
+                                result: Err(From::from(e)),
+                            }));
+                    }
+                };
+            }) {
+                let _ = self.event_tx.send(Event::ConnectionInfoPrepared(ConnectionInfoResult {
+                    result_token: result_token,
+                    result: Err(From::from(e)),
+                }));
+            }
         }
     }
 
@@ -398,7 +449,7 @@ impl Service {
     }
 }
 
-// Do not remvoe this as clones of mio-tx are give to IGD thread thus the drop of tx here alone is
+// Do not remove this as clones of mio-tx are give to IGD thread thus the drop of tx here alone is
 // not sufficient to trigger rx-error in core event loop for exit condition.
 impl Drop for Service {
     fn drop(&mut self) {
@@ -422,7 +473,6 @@ mod tests {
     use maidsafe_utilities;
     use maidsafe_utilities::thread::Joiner;
     use main::{Event, PrivConnectionInfo, PubConnectionInfo};
-
     use std::collections::{HashMap, HashSet, hash_map};
     use std::net::IpAddr;
     use std::str::FromStr;