+++ src/main/types.rs
@@ -33,15 +33,22 @@ pub struct PeerId(pub PublicKey);
 impl fmt::Debug for PeerId {
     fn fmt(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
         write!(formatter,
-               "PeerId({:02x}{:02x}..)",
+               "PeerId({:02x}{:02x}{:02x}{:02x}..)",
                (self.0).0[0],
-               (self.0).0[1])
+               (self.0).0[1],
+               (self.0).0[2],
+               (self.0).0[3])
     }
 }
 
 impl fmt::Display for PeerId {
     fn fmt(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        write!(formatter, "{:02x}{:02x}..", (self.0).0[0], (self.0).0[1])
+        write!(formatter,
+               "{:02x}{:02x}{:02x}{:02x}..",
+               (self.0).0[0],
+               (self.0).0[1],
+               (self.0).0[2],
+               (self.0).0[3])
     }
 }
 
@@ -85,7 +92,7 @@ pub struct PrivConnectionInfo {
     #[doc(hidden)]
     pub for_hole_punch: Vec<common::SocketAddr>,
     #[doc(hidden)]
-    pub hole_punch_socket: TcpBuilder,
+    pub hole_punch_socket: Option<TcpBuilder>,
 }
 
 impl PrivConnectionInfo {