+++ src/tests/mod.rs
@@ -19,10 +19,10 @@
 pub mod utils;
 
 pub use self::utils::{gen_config, get_event_sender, timebomb};
-use common::SocketAddr;
+
+use common::{CrustUser, SocketAddr};
 use main::{Config, Event, Service};
 use mio;
-
 use std::collections::HashSet;
 use std::net::SocketAddr as StdSocketAddr;
 use std::str::FromStr;
@@ -62,7 +62,7 @@ fn bootstrap_two_services_and_exchange_messages() {
     let (event_tx1, event_rx1) = get_event_sender();
     let mut service1 = unwrap!(Service::with_config(event_tx1, config1));
 
-    unwrap!(service1.start_bootstrap(HashSet::new()));
+    unwrap!(service1.start_bootstrap(HashSet::new(), CrustUser::Client));
 
     let peer_id0 = expect_event!(event_rx1, Event::BootstrapConnect(peer_id, _) => peer_id);
     assert_eq!(peer_id0, service0.id());
@@ -100,6 +100,9 @@ fn bootstrap_two_services_using_service_discovery() {
     let (event_tx1, event_rx1) = get_event_sender();
     let mut service1 = unwrap!(Service::with_config(event_tx1, config));
 
+    unwrap!(service1.start_listening_tcp());
+    let _ = expect_event!(event_rx1, Event::ListenerStarted(port) => port);
+
     service0.start_service_discovery();
     service0.set_service_discovery_listen(true);
     unwrap!(service0.start_listening_tcp());
@@ -107,7 +110,7 @@ fn bootstrap_two_services_using_service_discovery() {
     expect_event!(event_rx0, Event::ListenerStarted(_port));
 
     service1.start_service_discovery();
-    unwrap!(service1.start_bootstrap(HashSet::new()));
+    unwrap!(service1.start_bootstrap(HashSet::new(), CrustUser::Client));
 
     let peer_id0 = expect_event!(event_rx1, Event::BootstrapConnect(peer_id, _) => peer_id);
     assert_eq!(peer_id0, service0.id());
@@ -134,7 +137,10 @@ fn bootstrap_with_multiple_contact_endpoints() {
 
     let (event_tx1, event_rx1) = get_event_sender();
     let mut service1 = unwrap!(Service::with_config(event_tx1, config1));
-    unwrap!(service1.start_bootstrap(HashSet::new()));
+    unwrap!(service1.start_bootstrap(HashSet::new(), CrustUser::Client));
+
+    unwrap!(service1.start_listening_tcp());
+    let _ = expect_event!(event_rx1, Event::ListenerStarted(port) => port);
 
     let peer_id0 = expect_event!(event_rx1, Event::BootstrapConnect(peer_id, _) => peer_id);
     assert_eq!(peer_id0, service0.id());
@@ -163,7 +169,10 @@ fn bootstrap_with_blacklist() {
     let mut service1 = unwrap!(Service::with_config(event_tx1, config1));
     let mut blacklist = HashSet::new();
     blacklist.insert(*blacklisted_address);
-    unwrap!(service1.start_bootstrap(blacklist));
+    unwrap!(service1.start_bootstrap(blacklist, CrustUser::Client));
+
+    unwrap!(service1.start_listening_tcp());
+    let _ = expect_event!(event_rx1, Event::ListenerStarted(port) => port);
 
     let peer_id0 = expect_event!(event_rx1, Event::BootstrapConnect(peer_id, _) => peer_id);
     assert_eq!(peer_id0, service0.id());
@@ -195,7 +204,7 @@ fn bootstrap_fails_only_blacklisted_contact() {
 
     let mut blacklist = HashSet::new();
     blacklist.insert(*blacklisted_address);
-    unwrap!(service.start_bootstrap(blacklist));
+    unwrap!(service.start_bootstrap(blacklist, CrustUser::Client));
 
     expect_event!(event_rx, Event::BootstrapFailed);
 
@@ -213,7 +222,7 @@ fn bootstrap_fails_if_there_are_no_contacts() {
     let (event_tx, event_rx) = get_event_sender();
     let mut service = unwrap!(Service::with_config(event_tx, config));
 
-    unwrap!(service.start_bootstrap(HashSet::new()));
+    unwrap!(service.start_bootstrap(HashSet::new(), CrustUser::Client));
     expect_event!(event_rx, Event::BootstrapFailed);
 }
 
@@ -230,7 +239,7 @@ fn bootstrap_timeouts_if_there_are_only_invalid_contacts() {
     let (event_tx, event_rx) = get_event_sender();
     let mut service = unwrap!(Service::with_config(event_tx, config));
 
-    unwrap!(service.start_bootstrap(HashSet::new()));
+    unwrap!(service.start_bootstrap(HashSet::new(), CrustUser::Client));
     expect_event!(event_rx, Event::BootstrapFailed);
 }
 
@@ -248,7 +257,8 @@ fn drop_disconnects() {
 
     let (event_tx_1, event_rx_1) = get_event_sender();
     let mut service_1 = unwrap!(Service::with_config(event_tx_1, config_1));
-    unwrap!(service_1.start_bootstrap(HashSet::new()));
+
+    unwrap!(service_1.start_bootstrap(HashSet::new(), CrustUser::Client));
 
     let peer_id_0 = expect_event!(event_rx_1, Event::BootstrapConnect(peer_id, _) => peer_id);
     expect_event!(event_rx_0, Event::BootstrapAccept(_peer_id));
@@ -321,7 +331,7 @@ mod broken_peer {
                         let public_key = box_::gen_keypair().0;
                         unwrap!(self.0.write(poll,
                                              self.1,
-                                             Some((Message::BootstrapResponse(public_key),
+                                             Some((Message::BootstrapGranted(public_key),
                                                    0))));
                     }
                     Ok(Some(_)) | Ok(None) => (),
@@ -372,7 +382,7 @@ fn drop_peer_when_no_message_received_within_inactivity_period() {
     let (event_tx, event_rx) = get_event_sender();
     let mut service = unwrap!(Service::with_config(event_tx, config));
 
-    unwrap!(service.start_bootstrap(HashSet::new()));
+    unwrap!(service.start_bootstrap(HashSet::new(), CrustUser::Client));
     let peer_id = expect_event!(event_rx, Event::BootstrapConnect(peer_id, _) => peer_id);
 
     // The peer should drop after inactivity.
@@ -402,7 +412,7 @@ fn do_not_drop_peer_even_when_no_data_messages_are_exchanged_within_inactivity_p
     let (event_tx1, event_rx1) = get_event_sender();
     let mut service1 = unwrap!(Service::with_config(event_tx1, config1));
 
-    unwrap!(service1.start_bootstrap(HashSet::new()));
+    unwrap!(service1.start_bootstrap(HashSet::new(), CrustUser::Client));
     expect_event!(event_rx1, Event::BootstrapConnect(_peer_id, _));
     expect_event!(event_rx0, Event::BootstrapAccept(_peer_id));