+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CRUST - Change Log
 
+## [0.20.2]
+- Disable hole punching to limit concurrent socket usage.
+
+## [0.20.1]
+- Expose functionality to check if the connected peer is one of the hard-coded contacts from the config.
+
 ## [0.20.0]
 - Expose `Config` structure.
 - Update `rustfmt` options.